==> documentation/deeplearning/notebooks/01_data_exploration.ipynb <==
import pandas as pd
import numpy as np
import re

pd.set_option("display.max_columns", 85)
pd.set_option("display.max_rows", 85)

### Talkapalooza - TL BL SS 23

# Data exploration:

As a source of audio files for single word recordings, I downloaded the English language dataset from [Lingualibre's Languages Gallery](https://lingualibre.org/wiki/LinguaLibre:About). LinguaLibre is a project by Wikimédia France to collectively and collaboratively preserve and provide oral recordings of languages across the globe.

The English dataset consists of recordings provided by 109 unique speakers, amassing about 33 100 recordings covering 29 000 unique words. About 75% of recordings were provided by male speakers.

To have an overview over the pronunciations provided, as well as to know for which there are several recordings, some data preparation and analysis has to be performed.

## Exploring the dataset:

The dataset is downloaded as a folder with several subdirectories and files. Each subdirectory represents a user's recordings and file names describe the word recorded.

### You better walk that OS

import os

root = "../.source/lingualibre_en/"

recordings_df = pd.DataFrame(columns=['path', 'name'])
recordings_df.head()

Treat subdirs?

for path, subdirs, files in os.walk(root):
    for name in files:
        temp_df = pd.DataFrame([[path, name]], columns=['path', 'name'])
        recordings_df= pd.concat([recordings_df, temp_df])

recordings_df = recordings_df.reset_index()
recordings_df = recordings_df.drop('index', axis=1)

recordings_df.shape

recordings_df[::50].head(20)

### Clean up

Use regex to clean up path and file names. 

pattern_path = re.compile(".*/lingualibre_en/")
pattern_name = re.compile("\.ogg")

recordings_df['path'] =  [re.sub(pattern_path, '', str(x)) for x in recordings_df['path']]
recordings_df['name'] =  [x.lower() for x in recordings_df['name']]
recordings_df['name'] =  [re.sub(pattern_name, '', str(x)) for x in recordings_df['name']]

recordings_df['count'] = 1
recordings_df.head()

recordings_df = recordings_df.sort_values(by=['path', 'name'])

recordings_df = recordings_df.reset_index().drop('index', axis=1)
recordings_df.head()

### Insights

Unique users, find counts of unique words. Ideally we want to work with terms that have multiple recordings available.

users = recordings_df['path'].unique()
users.shape

Subdirectories are also present (in any case negligible data for us).

print(*users, sep=" | ")

recordings_df["name"].unique()

words = recordings_df["name"].value_counts()
words.shape

words = pd.DataFrame(words).reset_index().rename(columns={'index': 'word', 'name': 'count'})

words[::2500].head()

### Reshape

Reshape by looking at words, include count and usernames per word.

words_df = pd.DataFrame()

words_df[['word', 'count', 'user']] = (
    recordings_df.groupby(
        ['name']
    ).count().reset_index()
)

words_df.sort_values(by='count', ascending=False).head(400)

words_df['user'] = ''
words_df['user'] = (
    recordings_df.groupby(
        ['name'],
        as_index=False
    )['path']
)

words_df.head()

user_vals = words_df['user']
user_vals[:5]

for i, val in enumerate(user_vals):
    words_df.loc[i, 'user'] = ';'.join(user_vals[i][1].values)

words_df.sort_values(by='count', ascending=False).head(10)

words_df.iloc[24287]['user']

### Examine counts

filter_5 = (words_df['count'] >= 5)
filter_4 = (words_df['count'] >= 4)
filter_3 = (words_df['count'] >= 3)
filter_2 = (words_df['count'] >= 2)

print(words_df[filter_5].shape)
words_df[filter_5].sort_values(by='count', ascending=False)[::20].head(10)

print(words_df[filter_4].shape)
print(words_df[filter_3].shape)
print(words_df[filter_2].shape)
words_df[filter_2].sort_values(by='count', ascending=False)[::20].head(10)

## Cross Reference

### Read the csv used by Data Science team:

wordbank_df = pd.read_csv('../.source/wordbank_item_data.csv')

print(wordbank_df.shape)
wordbank_df.head()

wordbank_words = wordbank_df['item_definition'].unique()
print(wordbank_words[::30])

pattern_asterisk = re.compile('\*')
pattern_indicator = re.compile('\s\(.*\)')

wordbank_words = [re.sub(pattern_asterisk, '', str(x)) for x in wordbank_words]
wordbank_words = [re.sub(pattern_indicator, '', str(x)) for x in wordbank_words]

print(wordbank_words[::30])

filter_wordbank = (words_df['word'].isin(wordbank_words))
filter_wordbank_ext = ~(words_df['word'].isin(wordbank_words))

print(len(words_df[filter_wordbank]), 'words match in both')
words_df[filter_wordbank].head()

words_df[filter_wordbank & filter_2].sort_values(by='count').shape

#### Export

# words_df[filter_wordbank][['word', 'count']].to_csv('../output/wordbank_crossref.csv')

### Closer look at missing terms

wordbank_missing = list()
[wordbank_missing.append(w) for w in wordbank_words if w not in words_df[filter_wordbank]['word'].values]
print(len(wordbank_missing))

print(wordbank_missing)

filter_gotta = (words_df['word'] == 'gotta')

words_df[filter_gotta]

filter_gotto = (recordings_df['name'] == 'got to')

recordings_df[filter_gotto]

filter_got = (recordings_df['name'] == 'got')

recordings_df[filter_got]

#### Future notes

Make sure to include these by splitting strings at "/" and deleting any " to"'s?

If we want to have speaking exercises for these, that is - it is either slang ("gonna") or the bare infinitive form (vs. the more common/instinctual to-infinitive etc.).

wordbank_crossref_df = pd.read_csv('../output/wordbank_crossref.csv')

wordbank_crossref_df = wordbank_crossref_df.drop(columns="Unnamed: 0")

filter_3 = (wordbank_crossref_df['count'] >= 3)
filter_2 = (wordbank_crossref_df['count'] >= 2)

wordbank_crossref_df[filter_3][40:].head()

wordbank_crossref_df[filter_3][['word', 'count']].to_csv('../output/wordbank_crossref_filt3.csv')
