==> documentation/deeplearning/notebooks/02_data_preparation.ipynb <==
import pandas as pd
import numpy as np
import re

pd.set_option("display.max_columns", 85)
pd.set_option("display.max_rows", 85)

### Talkapalooza - TL BL SS 23

# Data preparation:

Based on previous notebook, reorder audio files in a way that they can be used for training a Deep Learning model.

Read the csv

words_df = pd.read_csv('../output/wordbank_crossref.csv')

word_list = words_df['word']

pattern_asterisk = re.compile('\*')
pattern_indicator = re.compile('\s\(.*\)')

word_list = [re.sub(pattern_asterisk, '', str(x)) for x in word_list]
word_list = [re.sub(pattern_indicator, '', str(x)) for x in word_list]

import os
import shutil
from pathlib import Path

root = "../.source/lingualibre_en/"
target = "../.source/labelled_audio/"

pattern_path = re.compile(".*/lingualibre_en/")
pattern_file = re.compile("\.ogg")

word_counter = 0
author_counter = 0

for path, subdirs, files in os.walk(root):
    for file in files:
        str_author = re.sub(pattern_path, '', str(path))
        str_author = re.sub(r'(?<=\w)\.', '_', str_author)
        str_author = re.sub(' ', '-', str_author)
        str_word = re.sub(pattern_file, '', str(file))
        str_file = f"{str_word}--{str_author}"

        if str_word not in word_list:
            continue

        src = os.path.abspath(Path(path, file))
        dst = os.path.abspath(Path(target) / str_word)
        dst = dst + '/'

        try:
            shutil.copyfile(src = src, dst = f"{dst}/{str_file}.ogg")
            author_counter += 1

        except:
            os.makedirs(os.path.dirname(dst), exist_ok=True)
            word_counter += 1
            shutil.copy(src = src, dst = f"{dst}/{str_file}.ogg")
            author_counter += 1
else:
    print(f'Total of {author_counter} recordings across {word_counter} labels.')


Alternative approach: all audio files in same directory.

for path, subdirs, files in os.walk(target):
    for file in files:
        src = os.path.abspath(Path(path, file))
        dst = os.path.abspath(Path("../.source/") / "labelled_data")
        dst = dst + '/'

        try:
            shutil.copyfile(src = src, dst = f"{dst}/{str(file)}")

        except:
            os.makedirs(os.path.dirname(dst), exist_ok=True)
            shutil.copy(src = src, dst = f"{dst}/{str(file)}")

## Verifying results:

pattern_target = re.compile(".*/labelled_audio/")

labels_df = pd.DataFrame()

for path, subdirs, files in os.walk(target):
    for file in files:
        str_label = re.sub(pattern_target, '', str(path))
        str_author = re.sub(pattern_file, '', str(file))
        temp_df = pd.DataFrame([[str_label, str_author]], columns=['label', 'author'])
        labels_df= pd.concat([labels_df, temp_df])

labels_df = labels_df.reset_index().drop(columns='index')

labels_df.head(10)
